+++ src/cebmf_torch/torch_cebnm/cash_solver.py
@@ -7,11 +7,11 @@ import torch.optim as optim
 from sklearn.preprocessing import StandardScaler
 from torch.utils.data import DataLoader, Dataset
 
-from cebmf_torch.utils.torch_distribution_operation import get_data_loglik_normal_torch
-from cebmf_torch.utils.torch_posterior import posterior_mean_norm
+from cebmf_torch.utils.distribution_operation import get_data_loglik_normal_torch
 
 # Import utils.py directly
-from cebmf_torch.utils.torch_utils_mix import autoselect_scales_mix_norm
+from cebmf_torch.utils.mixture import autoselect_scales_mix_norm
+from cebmf_torch.utils.posterior import posterior_mean_norm
 
 
 # Define dataset class that includes observation noise
