+++ src/cebmf_torch/torch_cebnm/emdn.py
@@ -4,8 +4,8 @@ import torch.optim as optim
 from sklearn.preprocessing import StandardScaler
 from torch.utils.data import DataLoader, Dataset
 
-from cebmf_torch.utils.torch_distribution_operation import get_data_loglik_normal_torch
-from cebmf_torch.utils.torch_posterior import posterior_mean_norm
+from cebmf_torch.utils.distribution_operation import get_data_loglik_normal_torch
+from cebmf_torch.utils.posterior import posterior_mean_norm
 
 
 # -------------------------
