+++ src/cebmf_torch/torch_ebnm/torch_ebnm_point_laplace.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 import torch
 from torch import Tensor
 
-from cebmf_torch.utils.torch_utils import (
+from cebmf_torch.utils.maths import (
     log_norm_pdf,  # kept for API parity (unused here)
     logPhi,
     my_e2truncnorm,
