+++ src/cebmf_torch/utils/device.py
