+++ src/cebmf_torch/utils/posterior.py
@@ -2,7 +2,7 @@ import math
 
 import torch
 
-from .torch_utils import my_e2truncnorm, my_etruncnorm
+from .maths import my_e2truncnorm, my_etruncnorm
 
 _LOG_SQRT_2PI = 0.5 * math.log(2.0 * math.pi)
