+++ tests/test_ash_posterior_mean_normal_torch.py
@@ -2,10 +2,9 @@ import numpy as np
 import torch
 
 from cebmf_torch.torch_ebnm.torch_ash import ash
-from cebmf_torch.utils.torch_distribution_operation import get_data_loglik_normal_torch
-from cebmf_torch.utils.torch_mix_opt import optimize_pi_logL
-from cebmf_torch.utils.torch_posterior import posterior_mean_norm
-from cebmf_torch.utils.torch_utils_mix import autoselect_scales_mix_norm
+from cebmf_torch.utils.distribution_operation import get_data_loglik_normal_torch
+from cebmf_torch.utils.mixture import autoselect_scales_mix_norm, optimize_pi_logL
+from cebmf_torch.utils.posterior import posterior_mean_norm
 
 
 def test_ash_loglik_and_scale():
