+++ tests/test_gaussian_gaussian_posterior_torch.py
@@ -1,8 +1,8 @@
 import numpy as np
 import torch
 
-from cebmf_torch.utils.torch_distribution_operation import get_data_loglik_normal_torch
-from cebmf_torch.utils.torch_utils_mix import autoselect_scales_mix_norm
+from cebmf_torch.utils.distribution_operation import get_data_loglik_normal_torch
+from cebmf_torch.utils.mixture import autoselect_scales_mix_norm
 
 
 def test_convolved_loglik_postmean():
